refactor: 기존 카트 아이템 계산 함수 제거 및 calculateItemTotal, getMaxApplicableDiscount 도메인 분리
